Remove the flag grow for the partition userdata
It seems that if the partition has the flag grow, it will not be updatable when
being packed into update.img and updated via RK flash tool. However, we need
this to be mounted on /boot and it is not a growable partition. Hence, we remove
this flag for it.

Change-Id: Iacb6e4cbf4ae844c285c1c6bb081a39e639aa252